api/requirements.txt
fastapi
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-decouple
deta
web3